third_party/WebKit/LayoutTests/http/tests/devtools/elements/elements-hide-html-comments.html

<!-- comment 1 -->
<html>
<head>
<script src="../../inspector/inspector-test.js"></script>
<script src="../../inspector/elements-test.js"></script>
<script>

function test() {
  ElementsTestRunner.nodeWithId('inspect', onNode);

  function onNode(node) {
    var commentNode = node.nextSibling;
    ElementsTestRunner.selectNode(commentNode).then(onNodeSelected);
  }

  function onNodeSelected() {
    TestRunner.addResult('HTML comments shown:');
    ElementsTestRunner.dumpElementsTree();
    Common.settingForTest('showHTMLComments').set(false);
    TestRunner.addResult('\nHTML comments hidden:');
    ElementsTestRunner.dumpElementsTree();
    TestRunner.completeTest();
  }
}

</script>
</head>
<!-- comment 2 -->
<body onload="runTest()">
<p>
Verifies show/hide HTML comments setting.
<span id="inspect"></span>
<!-- comment 3 -->
</p>
</body>
</html>
