third_party/WebKit/LayoutTests/http/tests/devtools/elements/elements-hide-html-comments-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Verifies show/hide HTML comments setting.

HTML comments shown:
  <!-- comment 1 -->
- <html>
    + <head>…</head>
      <!-- comment 2 -->
    - <body>
        - <p>
              <span id="inspect"></span>
              <!-- comment 3 -->
          </p>
      </body>
  </html>

HTML comments hidden:
- <html>
    + <head>…</head>
    - <body>
        + <p>…</p>
      </body>
  </html>
